+++ bot/commands/ability.js
@@ -32,8 +32,8 @@ The sort of ability that defines metagames. ex. Desolate Land, Shadow Tag`,
 		let ability = abilities[suffix];
 		if (ability === undefined){
 			let helper = utils.recognize(suffix, "ability");
-			msg.channel.sendMessage("```" + `Ability "${suffix}" not recognized. Did you mean "${helper[0]}"?` + "```");
-			ability = abilities[helper[0]];
+			msg.channel.sendMessage("```" + `Ability "${suffix}" not recognized. Did you mean "${helper.item.name}"?` + "```");
+			ability = helper.item;
 		}
 		
 		let sendMsg = `**${ability.name}**