+++ botEvents/ready.js
@@ -1,4 +1,6 @@
 module.exports = (bot, config) => {
-	bot.user.setAvatar(config.misc.avatar);
-	bot.user.setGame(config.misc.playing_game);
+	if (!bot.user.presence.game.name) {
+		bot.user.setAvatar(config.misc.avatar);
+		bot.user.setGame(config.misc.playing_game);
+	}
 }